added comment about CFG_SIMPLE_INT and x86_64 arcitectures

+++ examples/simple.c
@@ -7,7 +7,17 @@ int main(void)
     char *server = NULL;
     double delay = 1.356e-32;
     char *username = NULL;
-    int debug = 1;
+
+    /* Although the macro used to specify an integer option is called
+     * CFG_SIMPLE_INT(), it actually expects a long int. On a 64 bit system
+     * where ints are 32 bit and longs 64 bit (such as the x86-64 or amd64
+     * architectures), you will get weird effects if you use an int here.
+     *
+     * If you use the regular (non-"simple") options, ie CFG_INT() and use
+     * cfg_getint(), this is not a problem as the data types are implicitly
+     * cast.
+     */
+    long int debug = 1;
 
     cfg_opt_t opts[] = {
         CFG_SIMPLE_BOOL("verbose", &verbose),
